requirements.txt
Django==4.2.7
sqlparse==0.4.4
tzdata==2023.3
django-filter==25.1
drf-yasg==1.21.11
djangorestframework-simplejwt